pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gp-volatility-forecasting"
version = "0.1.0"
description = "Probabilistic stock volatility forecasting with Gaussian Processes and a Streamlit dashboard."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "gpytorch>=1.11",
  "numpy>=1.26",
  "pandas>=2.1",
  "pyarrow>=15.0",
  "PyYAML>=6.0",
  "scikit-learn>=1.4",
  "streamlit>=1.32",
  "torch>=2.2",
  "plotly>=5.19",
  "yfinance>=0.2.37",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools.packages.find]
include = ["src*"]
